change service file struct
