+++ .gitignore
@@ -1,6 +1,7 @@
 *.yml
-static/*
+lib/core/static/*
 go.sum
 .idea/*
 main
 .DS_Store
+*.log
