--- conf/cache.go
@@ -1,13 +1,0 @@
-package conf
-
-import (
-	"github.com/muesli/cache2go"
-)
-
-var Cache *cache2go.CacheTable
-
-func loadCache(){
-	if Cache == nil {
-		Cache = cache2go.Cache("myCache")
-	}
-}
