--- conf/crontab.go
@@ -1,16 +1,0 @@
-package conf
-
-import (
-	"fmt"
-	"github.com/whiteshtef/clockwork"
-)
-
-func Run() {
-	sched := clockwork.NewScheduler()
-	sched.Schedule().Every(1800).Seconds().Do(something)
-	sched.Run()
-}
-
-func something() {
-	fmt.Println("hello")
-}
