--- conf/source/eg.yml
@@ -1,14 +1,0 @@
-redis:
-  node: localhost:6379
-  db: 0
-  password:
-
-database:
-  driverName: mysql
-  url: root:@/imgo
-
-oss:
-  endPoint: your_endpoint
-  accessKeyId: your_accesskeyid
-  accessKeySecret: your_accesskeysecret
-  bucketName: your_bucket
