--- content/model/init.go
@@ -1,12 +1,0 @@
-package model
-
-import (
-	"fmt"
-	"stouch_server/conf"
-)
-
-func init(){
-	if err := conf.Orm.Sync2(new(Topic)); err != nil {
-		fmt.Println(err)
-	}
-}
