+++ main.go
@@ -8,12 +8,13 @@ import (
 	"github.com/kataras/iris/middleware/logger"
 	"github.com/kataras/iris/middleware/recover"
 	"github.com/kataras/iris/mvc"
-	appconfCt "stouch_server/appconf/controller"
-	authCt "stouch_server/auth/controller"
-	"stouch_server/conf"
-	contentCt "stouch_server/content/controller"
-	storageCt "stouch_server/storage/controller"
-	bookCt "stouch_server/websock/controller"
+	"net/http"
+	appconfCt "stouch_server/src/appconf/controller"
+	authCt "stouch_server/src/auth/controller"
+	contentCt "stouch_server/src/content/controller"
+	"stouch_server/src/core"
+	storageCt "stouch_server/src/storage/controller"
+	bookCt "stouch_server/src/websock/controller"
 )
 
 func newApp() *iris.Application {
@@ -23,7 +24,7 @@ func newApp() *iris.Application {
 	// 解决跨域问题
 	allowAllOrigins := func(ctx iris.Context) {
 		ctx.Header("Access-Control-Allow-Origin", "*")                                         // <- HERE
-		ctx.Header("Access-Control-Allow-Headers", "ticket, Content-Type, app, authorization") // <- HERE
+		ctx.Header("Access-Control-Allow-Headers", "ticket, Content-Type, src, authorization") // <- HERE
 		ctx.Header("Access-Control-Allow-Methods", "DELETE, GET, OPTIONS, PATCH, POST, PUT")
 		if ctx.Method() != "OPTIONS" {
 			ctx.Next()
@@ -31,10 +32,7 @@ func newApp() *iris.Application {
 	}
 	app.Use(allowAllOrigins)
 	app.AllowMethods(iris.MethodOptions)
-
-	app.Get("/", func(ctx iris.Context) { ctx.Redirect("http://airport.xiaorere.com/index.html") })
 	app.StaticWeb("/web", "./static")
-	mvc.New(app.Party("/user")).Handle(new(authCt.UserController))
 	mvc.New(app.Party("/content")).Handle(new(contentCt.ContentController))
 	mvc.New(app.Party("/storage/token")).Handle(new(storageCt.StorageTokenController))
 	mvc.New(app.Party("/storage/picture")).Handle(new(storageCt.PictureController))
@@ -43,9 +41,19 @@ func newApp() *iris.Application {
 }
 
 func main() {
+	// 禁用控制台颜色，将日志写入文件时不需要控制台颜色。
+	gin.DisableConsoleColor()
+
 	r := gin.Default()
+	r.GET("/", func(c *gin.Context) {
+		c.Redirect(http.StatusTemporaryRedirect, "http://airport.xiaorere.com/index.html")
+	})
+	r.Static("/static", "./static")
 	appconfCt.AddRoutes(r.Group("/appconf"))
-	go conf.Run()
+	authCt.AddRoutes(r.Group("/auth"))
+
+	// 定时任务
+	go core.Run()
 	err := r.Run()
 	if err != nil {
 		fmt.Println(err)
