+++ resources/static/index.html
@@ -1,0 +1,1 @@
+<h>This is first static page!!!</h>
