+++ src/auth/controller/route.go
@@ -1,0 +1,9 @@
+package controller
+
+import (
+"github.com/gin-gonic/gin"
+)
+
+func AddRoutes(rg *gin.RouterGroup) {
+	rg.GET("/user/:id", GetBy)
+}
